scx_flow: 3.0.3-cherrypick-step2 — modular code split and README rewrite
Split the monolithic main.bpf.c (615 lines) into 7 focused modules,
all compiled as one translation unit via #include from main.bpf.c:

  main.bpf.c         (136 lines)  — maps, volatiles, init/exit, ops table
  helpers.bpf.c      (158 lines)  — utility functions and macros
  select_cpu.bpf.c   (77 lines)   — CPU selection
  enqueue.bpf.c      (120 lines)  — enqueue routing (pinned -> vtime)
  dispatch.bpf.c     (34 lines)   — DSQ hierarchy dispatch
  budget.bpf.c       (63 lines)   — budget lifecycle (runnable/running/stopping)
  task.bpf.c         (56 lines)   — task lifecycle (init/enable/exit/yield/cpu_release)

Each module has a single clear responsibility.  No behavioural changes
from the 3.0.0 base (commit 2e1e3c84).  No hybrid CPU awareness code
included — the Intel/AMD P-core/E-core logic from 3.0.3-step1 is
excluded to avoid the regression identified in benchmark comparisons.

Rewrite README in the concise format introduced in the 3.1.0 branch:
Overview, Typical Use Case, Production Readiness, and Validation.
