+++ scheds/experimental/scx_flow/src/bpf/budget.bpf.c
@@ -1,0 +1,64 @@
+/* Budget lifecycle tracking — included by main.bpf.c via #include */
+
+void BPF_STRUCT_OPS(flow_runnable, struct task_struct *p, u64 enq_flags)
+{
+	struct task_ctx *tctx;
+	u64 now;
+
+	tctx = lookup_task_ctx(p);
+	if (!tctx)
+		return;
+
+	now = bpf_ktime_get_ns();
+	if (tctx->sleep_started_at && now > tctx->sleep_started_at)
+		FLOW_CPUSTAT_INC(lookup_cpu_state(), runnable_wakeups);
+	update_budget_on_wakeup(p, tctx, now);
+}
+
+void BPF_STRUCT_OPS(flow_running, struct task_struct *p)
+{
+	struct task_ctx *tctx;
+	s32 current_cpu;
+	u64 now;
+
+	tctx = lookup_task_ctx(p);
+	current_cpu = bpf_get_smp_processor_id();
+	now = bpf_ktime_get_ns();
+	if (tctx) {
+		if (tctx->last_cpu >= 0 && tctx->last_cpu != current_cpu)
+			FLOW_CPUSTAT_INC(lookup_cpu_state(), cpu_migrations);
+		tctx->last_cpu = current_cpu;
+		tctx->last_run_at = now;
+		tctx->first_run = false;
+	}
+
+	__sync_fetch_and_add(&nr_running, 1);
+}
+
+void BPF_STRUCT_OPS(flow_stopping, struct task_struct *p, bool runnable)
+{
+	struct task_ctx *tctx;
+	u64 now;
+	u64 runtime_ns = 0;
+
+	tctx = lookup_task_ctx(p);
+	now = bpf_ktime_get_ns();
+
+	if (tctx) {
+		if (tctx->last_run_at && now > tctx->last_run_at)
+			runtime_ns = now - tctx->last_run_at;
+
+		if (tctx->budget_ns > 0 &&
+		    tctx->budget_ns - (s64)runtime_ns <= 0)
+			FLOW_CPUSTAT_INC(lookup_cpu_state(), budget_exhaustions);
+
+		tctx->budget_ns = clamp_budget(tctx->budget_ns - (s64)runtime_ns);
+		tctx->last_run_at = 0;
+		tctx->sleep_started_at = runnable ? 0 : now;
+		if (!runnable)
+			clear_wake_target(tctx);
+	}
+
+	__sync_fetch_and_add(&total_runtime, runtime_ns);
+	__sync_fetch_and_sub(&nr_running, 1);
+}
