+++ scheds/experimental/scx_flow/src/bpf/select_cpu.bpf.c
@@ -1,0 +1,52 @@
+/* CPU selection — included by main.bpf.c via #include */
+
+s32 BPF_STRUCT_OPS(flow_select_cpu, struct task_struct *p, s32 prev_cpu, u64 wake_flags)
+{
+	struct task_ctx *tctx;
+	bool is_idle = false;
+	s32 cpu;
+	s32 preferred_cpu;
+	s32 this_cpu = bpf_get_smp_processor_id();
+	bool non_migratable = is_non_migratable(p);
+	bool is_this_cpu_allowed = bpf_cpumask_test_cpu(this_cpu, p->cpus_ptr);
+
+	tctx = lookup_task_ctx(p);
+	if (tctx) {
+		if (tctx->sleep_started_at)
+			update_budget_on_wakeup(p, tctx, bpf_ktime_get_ns());
+		clear_wake_target(tctx);
+	}
+
+	if (!bpf_cpumask_test_cpu(prev_cpu, p->cpus_ptr))
+		prev_cpu = is_this_cpu_allowed ? this_cpu : bpf_cpumask_first(p->cpus_ptr);
+
+	preferred_cpu = prev_cpu;
+	if (!non_migratable && tctx && tctx->last_cpu >= 0 &&
+	    bpf_cpumask_test_cpu(tctx->last_cpu, p->cpus_ptr))
+		preferred_cpu = tctx->last_cpu;
+
+	if (non_migratable) {
+		cpu = preferred_cpu;
+		is_idle = scx_bpf_test_and_clear_cpu_idle(preferred_cpu);
+	} else if (tctx && tctx->first_run) {
+		cpu = scx_bpf_pick_idle_cpu(p->cpus_ptr, FLOW_PICK_IDLE_CORE);
+		if (cpu < 0)
+			cpu = scx_bpf_pick_idle_cpu(p->cpus_ptr, 0);
+		if (cpu >= 0)
+			is_idle = true;
+		else
+			cpu = scx_bpf_select_cpu_dfl(p, preferred_cpu, wake_flags, &is_idle);
+	} else {
+		cpu = scx_bpf_select_cpu_dfl(p, preferred_cpu, wake_flags, &is_idle);
+	}
+
+	if (tctx) {
+		tctx->wake_cpu = cpu >= 0 ? cpu : preferred_cpu;
+		tctx->wake_cpu_idle = is_idle;
+		tctx->wake_cpu_valid =
+			tctx->wake_cpu >= 0 &&
+			bpf_cpumask_test_cpu(tctx->wake_cpu, p->cpus_ptr);
+	}
+
+	return cpu >= 0 ? cpu : preferred_cpu;
+}
