+++ scheds/experimental/scx_flow/src/bpf/task.bpf.c
@@ -1,0 +1,57 @@
+/* Task lifecycle — included by main.bpf.c via #include */
+
+s32 BPF_STRUCT_OPS_SLEEPABLE(flow_init_task, struct task_struct *p,
+			     struct scx_init_task_args *args)
+{
+	struct task_ctx *tctx;
+	u64 now;
+
+	tctx = alloc_task_ctx(p);
+	if (!tctx)
+		return -ENOMEM;
+
+	now = bpf_ktime_get_ns();
+	reset_task_ctx(tctx, now, true);
+	__sync_fetch_and_add(&init_task_events, 1);
+
+	return 0;
+}
+
+void BPF_STRUCT_OPS(flow_enable, struct task_struct *p)
+{
+	struct task_ctx *tctx;
+	bool sleeping;
+
+	tctx = lookup_task_ctx(p);
+	if (!tctx)
+		return;
+
+	sleeping = !scx_bpf_task_running(p);
+	reset_task_ctx(tctx, bpf_ktime_get_ns(), sleeping);
+	__sync_fetch_and_add(&enable_events, 1);
+}
+
+bool BPF_STRUCT_OPS(flow_yield, struct task_struct *from, struct task_struct *to)
+{
+	from->scx.slice = FLOW_SLICE_MIN_NS;
+	return false;
+}
+
+void BPF_STRUCT_OPS(flow_cpu_release, s32 cpu, struct scx_cpu_release_args *args)
+{
+	scx_bpf_reenqueue_local();
+	__sync_fetch_and_add(&cpu_release_reenqueues, 1);
+}
+
+void BPF_STRUCT_OPS(flow_exit_task, struct task_struct *p,
+		    struct scx_exit_task_args *args)
+{
+	struct task_ctx *tctx;
+
+	tctx = lookup_task_ctx(p);
+	if (!tctx)
+		return;
+
+	reset_task_ctx(tctx, 0, false);
+	__sync_fetch_and_add(&exit_task_events, 1);
+}
